requirements.txt
Flask==3.1.2
tensorflow==2.20.0
numpy
opencv-python-headless
Pillow